Move TestEntropyh20Func to utils test
